+++ htdocs/examples/lqdn-gafam-campaign/index.js
@@ -1,1 +1,2 @@
 require('./poster.js');
+require('./chooser.js');