+++ CHANGES.rst
@@ -5,9 +5,13 @@ PosterKit-O-Mat CHANGES
 
 in progress
 -----------
-- lqdn-gafam-campaign: Obtain "lang" and "name" variables from query parameters,
-  e.g. https://examples.posterkit.net/lqdn-gafam-campaign/poster.html?lang=fr&name=google
-- lqdn-gafam-campaign: Apply custom layout settings based on poster name
+- lqdn-gafam-campaign
+
+    - Obtain "lang" and "name" variables from query parameters,
+      e.g. https://examples.posterkit.net/lqdn-gafam-campaign/poster.html?lang=fr&name=google
+    - Apply custom layout settings based on poster name
+    - Add matrix-based chooser for language vs. name,
+      see. https://examples.posterkit.net/lqdn-gafam-campaign/chooser.html
 
 2018-01-15 0.0.0
 ----------------